+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## [Unreleased]
 ### Fixed
 - Fixed multiple return types in stubs [#2283](https://github.com/zephir-lang/zephir/issues/2283)
+### Changed
+- Removed `.zep` from stubs filenames [#2273](https://github.com/zephir-lang/zephir/issues/2273)
 
 ## [0.14.0] - 2021-09-18
 ### Added